requirements.txt
fastapi
uvicorn
websockets
python-dotenv
google-generativeai
faster-whisper
edge-tts
pydantic
numpy
